partie I + makefile & libft.h

+++ libft/ft_strlcpy.c
@@ -10,7 +10,9 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-unsigned int	ft_strlcpy(char *dest, char *src, unsigned int size)
+#include "libft.h"
+
+size_t	ft_strlcpy(char *dest, char *src, unsigned int size)
 {
 	unsigned int	l;
 	unsigned int	i;
